# Fix Latency Tracking Issue