--- a/libs/server/Resp/Parser/RespCommand.cs
+++ b/libs/server/Resp/Parser/RespCommand.cs
@@ -1993,9 +1993,13 @@ namespace Garnet.server
             }
             endReadHead = (int)(ptr - recvBufferPtr);
 
-            // Classify command as non data command
+            // Classify command as slow to avoid accounting for its latency in NET_RS
             if (storeWrapper.serverOptions.LatencyMonitor)
-                containsSlowCommand |= !cmd.IsDataCommand();
+            {
+                var normCmd = cmd.NormalizeForACLs();
+                _ = RespCommandsInfo.TryFastGetRespCommandInfo(normCmd, out var commandInfo);
+                containsSlowCommand |= commandInfo == null || !commandInfo.Flags.HasFlag(RespCommandFlags.Fast);
+            }
 
             if (storeWrapper.appendOnlyFile != null && storeWrapper.serverOptions.WaitForCommit)
                 HandleAofCommitMode(cmd);